requirements.txt
Django==4.2.3
PyYAML==6.0
asgiref==3.7.2
attrs==23.1.0
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
drf-spectacular==0.26.3
inflection==0.5.1
jsonschema==4.18.2
jsonschema-specifications==2023.6.1
pip==23.1.2
psycopg2-binary==2.9.6
pytz==2023.3
referencing==0.29.1
rpds-py==0.8.10
setuptools==67.7.2
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
wheel==0.40.0